+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+### WIP
+
+* Methods in `FutureOps` now require an implicit `ExecutionContext` (before, the global EC was used)
+* `Fetch` builders now require an implicit `ExecutionContext` (before, the global EC was used) 
+
 ### 0.14.3
 
 * Bugfix: in the WebSocket client, a check for the "connected" status was incomplete - https://github.com/tulz-app/laminext/pull/108 (@lavrov)